requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.4
pypdf>=4.2.0
python-docx>=1.1.0
python-dotenv>=1.0.1
